backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Werkzeug==3.0.1
SQLAlchemy==2.0.25
Pillow==10.2.0
reportlab==4.1.0
python-dotenv==1.0.0
requests==2.31.0
pyserial==3.5
psutil==5.9.8
colorlog==6.8.2
python-dateutil==2.8.2
exifread==3.0.0
qrcode==7.4.2
fpdf2==2.7.8
